LPS-155445 Changes imports and forders structure
