+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/home/components/SearchHeader/components/SearchBar/_SearchBar.scss
@@ -1,0 +1,14 @@
+.cp-search-project {
+	padding: 1.125rem var(--form-input-padding-x) 1.125rem 4rem;
+	width: 553px;
+
+	&:hover {
+		border-color: var(--color-action-primary-hover);
+	}
+
+	& + .lexicon-icon {
+		bottom: 25.63px;
+		font-size: 1.421rem;
+		left: 28.63px;
+	}
+}
